feat: implement scrolling text component and karaoke scroll progress hook for lyrics display
