+++ src/renderer/styles/lyrics/lyrics.css
@@ -67,6 +67,7 @@
 
 .lyrics-lines-wrapper {
   display: flex;
+  flex: 1;
   flex-direction: column;
   align-items: flex-end;
   gap: 2px;
@@ -117,6 +118,54 @@
   background-clip: text;
   -webkit-text-fill-color: transparent;
   color: transparent;
+}
+
+/* 溢出文本横向滚动 — 复用 ScrollingText 组件 */
+.scroll-text {
+  width: 100%;
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: clip;
+  white-space: nowrap;
+  text-align: right;
+}
+
+.scroll-text-content {
+  display: inline-block;
+  min-width: max-content;
+  transform: translateX(0);
+}
+
+.scroll-text.is-overflowing {
+  text-align: left;
+}
+
+.scroll-text.is-overflowing .scroll-text-content {
+  animation: scroll-text-scroll var(--scroll-text-duration) ease-in-out 1s infinite alternate;
+}
+
+.scroll-text.is-progress-driven .scroll-text-content {
+  animation: none;
+  transform: translateX(calc(-1 * var(--scroll-text-offset)));
+  transition: transform 0.12s linear;
+}
+
+@keyframes scroll-text-scroll {
+  from {
+    transform: translateX(0);
+  }
+
+  to {
+    transform: translateX(calc(-1 * var(--scroll-text-distance)));
+  }
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .scroll-text.is-overflowing .scroll-text-content {
+    animation: none;
+    transform: translateX(0);
+    transition: none;
+  }
 }
 
 .lyrics-loading {
